+++ docs/source/ttnn/ttnn/dependencies/tt_lib.rst
@@ -824,8 +824,6 @@ Backward Operations
 
 .. autofunction:: tt_lib.tensor.gt_bw
 
-.. autofunction:: tt_lib.tensor.relu_bw
-
 .. autofunction:: tt_lib.tensor.ne_bw
 
 .. autofunction:: tt_lib.tensor.clamp_max_bw
