+++ docs/source/ttnn/ttnn/ttnn/relu_bw.rst
@@ -1,0 +1,6 @@
+.. _ttnn.relu_bw:
+
+ttnn.relu_bw
+#################
+
+.. autofunction:: ttnn.relu_bw
