+++ tests/tt_eager/python_api_testing/unit_testing/backward_ops/test_backward_relu.py
@@ -4,7 +4,7 @@
 
 import torch
 import pytest
-import tt_lib
+import ttnn
 from tests.tt_eager.python_api_testing.unit_testing.backward_ops.utility_funcs import compare_pcc, data_gen_with_range
 
 
@@ -23,7 +23,7 @@ def test_bw_relu(input_shapes, device):
 
     pyt_y = torch.relu(in_data)
 
-    tt_output_tensor_on_device = tt_lib.tensor.relu_bw(grad_tensor, input_tensor)
+    tt_output_tensor_on_device = ttnn.relu_bw(grad_tensor, input_tensor)
 
     in_data.retain_grad()
