+++ ttnn/cpp/ttnn/operations/eltwise/unary_backward/device/unary_backward_op.cpp
@@ -220,6 +220,13 @@ std::vector<Tensor> _neg_bw(const Tensor& grad, const Tensor& input, const Memor
     return grad_tensor;
 }
 
+std::vector<Tensor> _relu_bw(const Tensor& grad, const Tensor& input, const MemoryConfig& output_mem_config) {
+    std::vector<Tensor> grad_tensor;
+    Tensor result = ttnn::multiply(ttnn::gtz(input, output_mem_config), grad, std::nullopt, output_mem_config);
+    grad_tensor.emplace_back(result);
+    return grad_tensor;
+}
+
 std::function<std::vector<ttnn::Tensor>(const Tensor&, const Tensor&, const MemoryConfig&)> UnaryBackwardFunction::get_function_type1(UnaryBackwardOpType OpType){
     switch (OpType) {
         case UnaryBackwardOpType::ASSIGN_BW:
@@ -246,6 +253,8 @@ std::function<std::vector<ttnn::Tensor>(const Tensor&, const Tensor&, const Memo
             return _rsqrt_bw;
         case UnaryBackwardOpType::NEG_BW:
             return _neg_bw;
+        case UnaryBackwardOpType::RELU_BW:
+            return _relu_bw;
         default:
             TT_ASSERT(false && "Undefined op type");
             return 0;
